test(refactor): consolidate duplicated Mongo and S3 test fakes (#376)
- Move FakeCursor, _matches, FakeCollection, FakeDatabase, and FakeStorage from
  tests/api/conftest.py, tests/integration/conftest.py, and
  tests/infrastructure/test_ingestion_persistence.py into tests/conftest.py.
- Extend the shared matcher to support $or, $gt, $gte, $lt, $lte, $exists,
  $in, $ne, and $regex, matching the queries used across the test suites.
- Update all three consumers to import the shared fakes instead of maintaining
  local copies.
- Keep test-specific subclasses (StrictStorage, AccessDeniedStorage) in the
  persistence test file.
